// src/models/raw-book.ts
export interface RawBook {
    ['﻿"Titre de la série"']: string;
    Type: string;
    Collection: string;
    Catégory: string;
    ['Titre de l\'album']: string;
    EAN: string;
    Tome: string;
    ['Date de publication']: string;
    Editeur: string;
    Auteurs: string;
    Prix: string;
    ['Date d\'ajout']: string;
    ['Dans ma collection']: string;
    Lu: string;
    Dédicacé: string;
    ['Edition originale']: string;
    ['Prété à']: string;
    pageCount: number
}